+++ CHANGES.rst
@@ -15,6 +15,8 @@ Unreleased (1.2.25) (XXXX-XX-XX)
 
 - Refactor drawLines to drawCircles.
 
+- Add zoom to data bounds for events.
+
 - Added maximum number of timeseries events to prevent browser running out of
   memory.